+++ package.json
@@ -4,14 +4,10 @@
   "author": "Stacey-Gammon",
   "dependencies": {
     "glob": "^7.1.6",
+    "gulp": "4.0.2",
     "gulp-cli": "^2.2.0",
     "gulp-debug": "^4.0.0",
     "gulp-watch": "^5.0.1",
-    "js-yaml": "^3.13.1",
-    "gulp": "4.0.2"
-  },
-  "scripts": {
-    "start": "node ./scripts/start",
-    "kbn": "node ./scripts/kbn"
+    "js-yaml": "^3.13.1"
   }
 }
