--- plugins/greeting_cards_story/greeting_cards_v1/kibana.json
@@ -1,10 +1,0 @@
-{
-  "id": "greetingCards",
-  "version": "0.0.1",
-  "kibanaVersion": "kibana",
-  "configPath": ["greeting_cards"],
-  "server": false,
-  "ui": true,
-  "requiredPlugins": [],
-  "optionalPlugins": []
-}
