--- plugins/greeting_cards_story/greeting_cards_v1/tsconfig.json
@@ -1,15 +1,0 @@
-{
-  "extends": "../../tsconfig.json",
-  "compilerOptions": {
-    "outDir": "./target",
-    "skipLibCheck": true
-  },
-  "include": [
-    "index.ts",
-    "public/**/*.ts",
-    "public/**/*.tsx",
-    "server/**/*.ts",
-    "../../typings/**/*",
-  ],
-  "exclude": []
-}
