+++ plugins/greeting_cards_v1/public/index.ts
@@ -1,0 +1,3 @@
+import { GreetingCardsPlugin } from './plugin';
+
+export const plugin = () => new GreetingCardsPlugin();
