+++ plugins/greeting_cards_v1/public/types.ts
@@ -1,0 +1,8 @@
+export const BIRTHDAY = 'birthday';
+export const GET_WELL = 'getWellSoon';
+
+export interface GreetingCard {
+  message: string;
+  to: string;
+  from: string;
+}
