--- scripts/start.js
@@ -1,3 +1,0 @@
-require('../gulpfile');
-
-gulp.series('default');
